requirements.txt
config==0.3.9
dj-database-url==0.5.0
Django==2.2.8
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
